# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEROAD",
        "description": "Api documentation",
        "contact": {
            "email": "peramkat@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://theme.zdassets.com/theme_assets/9115960/ef5800cc529889d180b05b57e40dd50e5c7adb73.png"
        }
    },
    "paths": {
...
        "/api/v1/travels": {
            "get": {
                "tags": [
                    "Api v1 - Travels"
                ],
                "summary": "Get all travels",
                "description": "Retrieve a list of all travels",
                "operationId": "bfe3c61041966f70bfcafc8019bd7fd4",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/travel": {
            "post": {
                "tags": [
                    "Api v1 - Travels"
                ],
                "summary": "Create a new travel",
                "description": "Store a newly created travel in the database",
                "operationId": "b9effdf8c24f61ad7f0b77849f110828",
                "requestBody": {
                    "description": "Travel details",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "isPublic": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "new travel"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "new-travel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "content of the first travel"
                                    },
                                    "numberOfDays": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "nature": {
                                        "type": "integer",
                                        "example": "20"
                                    },
                                    "relax": {
                                        "type": "integer",
                                        "example": "30"
                                    },
                                    "history": {
                                        "type": "integer",
                                        "example": "40"
                                    },
                                    "culture": {
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "party": {
                                        "type": "integer",
                                        "example": "60"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Travel created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Travel created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/travels/{travel:slug}": {
            "put": {
                "tags": [
                    "Api v1 - Travels"
                ],
                "summary": "Update a travel",
                "description": "Update an existing travel in the database base on the give slug",
                "operationId": "19f184c137c5db272277078b85da4a36",
                "parameters": [
                    {
                        "name": "travel:slug",
                        "in": "path",
                        "description": "slug of the travel to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        },
                        "example": "united-arab-emirates"
                    }
                ],
                "requestBody": {
                    "description": "Travel details",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "isPublic",
                                    "name",
                                    "numberOfDays",
                                    "description"
                                ],
                                "properties": {
                                    "isPublic": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "new united arab emirates"
                                    },
                                    "numberOfDays": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "5"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "new content"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Travel updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Travel updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}